+++ ntex-net/CHANGES.md
@@ -2,7 +2,9 @@
 
 ## [2.5.20] - 2025-05-19
 
-* Do not use zc send for small buffers
+* iour: Handle POLLRDHUP event
+
+* iour: Do not use zc send for small buffers
 
 ## [2.5.19] - 2025-05-15